+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/block/fire_behaviors/EnderPearlBehavior.java
@@ -7,6 +7,7 @@ import net.minecraft.core.BlockSourceImpl;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.projectile.Projectile;
+import net.minecraft.world.entity.projectile.ThrownEnderpearl;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.phys.Vec3;
 import org.jetbrains.annotations.Nullable;
@@ -17,9 +18,10 @@ public class EnderPearlBehavior extends GenericProjectileBehavior {
     public boolean fire(ItemStack stack, ServerLevel level, Vec3 pos, Vec3 facing,
                         float power, float drag, int inaccuracy, @Nullable Player owner) {
         BlockSource source = new BlockSourceImpl(level, BlockPos.containing(pos));
-        Projectile pearl = PearlMarker.createPearlToDispenseAndPlaceMarker(source, pos);
+        Projectile pearl = owner != null ? new ThrownEnderpearl(level, owner) :
+                PearlMarker.createPearlToDispenseAndPlaceMarker(source, pos);
 
-        pearl.shoot(facing.x, facing.y, facing.z, -drag * power, inaccuracy);
+        pearl.shoot(facing.x, facing.y, facing.z, drag * power, inaccuracy);
 
         pearl.setPos(pos);
         level.addFreshEntity(pearl);